feat: 코인 작성 상세화면 구현

+++ 9th_muyaho_iOS/models/WriteStockReqeust.swift
@@ -9,7 +9,7 @@ struct WriteStockRequest {
     
     let stockId: Int
     let purchasePrice: Double
-    let quantity: Int
+    let quantity: Double
     let currencyType: CurrencyType
     let purchaseTotalPrice: Double
